requirements.txt
accelerate==0.15.0
bitsandbytes
diffusers==0.11.1
huggingface_hub
numpy==1.21
torchvision
transformers>=4.25.1
